restructure and enhance MultiSelectComboBox validation tests
