+++ vaadin-combo-box-flow-parent/vaadin-combo-box-flow/src/test/java/com/vaadin/flow/component/combobox/validation/MultiSelectComboBoxBinderValidationTest.java
@@ -1,5 +1,6 @@
-package com.vaadin.flow.component.combobox;
+package com.vaadin.flow.component.combobox.validation;
 
+import com.vaadin.flow.component.combobox.MultiSelectComboBox;
 import com.vaadin.flow.data.binder.Binder;
 import com.vaadin.flow.data.binder.ValidationResult;
 import com.vaadin.flow.data.binder.Validator;
